+++ SFMathConsole/SFMathConsole.h
@@ -9,10 +9,10 @@
 extern "C"
 {
     //DLL Import of FFT_PSD from DLL
-    __declspec(dllimport) void __stdcall FFT_PSD(float *data, unsigned int data_len, bool use_hanning, float *out_data, int*out_data_len);
-    __declspec(dllimport) void __stdcall FFT_PSD_windows(float *data, unsigned int data_len, unsigned int window_size, bool use_hanning, float *out_data, int*out_data_len);
+    __declspec(dllimport) void __stdcall FFT_PSD(double *data, unsigned int data_len, bool use_hanning, double *out_data, int*out_data_len);
+    __declspec(dllimport) void __stdcall FFT_PSD_windows(double *data, unsigned int data_len, unsigned int window_size, bool use_hanning, double *out_data, int*out_data_len);
     __declspec(dllimport) void __stdcall SFMath_Hello();
-    __declspec(dllimport) void __stdcall CSpline_Interp(float *X, float*Y, unsigned int N, float *intXs, float *predYs, unsigned int NPreds);
+    __declspec(dllimport) void __stdcall CSpline_Interp(double *X, double*Y, unsigned int N, double *intXs, double *predYs, unsigned int NPreds);
 };
 
 //Test functions
